feat(endpoints): criando endpoint para criação de usuário, upload do roteiro e protótipo de mudança de status

+++ src/main/java/br/com/riume/backendeureka/model/Status.java
@@ -13,7 +13,7 @@ public class Status {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "status_id")
-    private Integer statusId;
+    private Long statusId;
 
     @Column(name = "status_name", nullable = false, unique = true)
     private String statusName;
